requirements-dev.txt
attrs >=17.4
black >=22.12.0
bump-my-version
compas_invocations2
compas_viewer
invoke >=0.14
ruff
sphinx_compas2_theme
twine
wheel
scikit-build-core
nanobind
cmake
build